separation ExecutionPlan from DataSource
